Revert "funktioner2"
This reverts commit e6e679c456f962337501e18c06f10be489ec7b48.

+++ PasswordManagerProjekt/Utility/VaultManager.cs
@@ -21,21 +21,21 @@ namespace PwM_UI.Utility
                 string pathToVault = Path.Combine(vaultDirectory, $"{vaultName}.x");
                 if (File.Exists(pathToVault))
                 {
-                    Notification.ShowNotificationInfo("orange", $"Warning CreateVault: Vault {vaultName} already exist!");
+                    Notification.ShowNotificationInfo("orange", $"Vault {vaultName} already exist!");
                     Globals.vaultChecks = true;
                     return;
                 }
 
                 if (vaultName.Length < 3)
                 {
-                    Notification.ShowNotificationInfo("orange", "Warning CreateVault: Vault name must be at least 3 characters long.");
+                    Notification.ShowNotificationInfo("orange", "Vault name must be at least 3 characters long.");
                     Globals.vaultChecks = true;
                     return;
                 }
 
                 if (!PasswordService.ValidatePassword(confirmPassword))
                 {
-                    Notification.ShowNotificationInfo("orange", "Warning CreateVault: Password must be at least 12 characters and include at least 1 upper case, 1 lower case, 1 digit, 1 special character and no white space!");
+                    Notification.ShowNotificationInfo("orange", "Password must be at least 12 characters and include at least 1 upper case, 1 lower case, 1 digit, 1 special character and no white space!");
                     Globals.vaultChecks = true;
                     return;
                 }
@@ -51,29 +51,41 @@ namespace PwM_UI.Utility
             }
         }
 
+
+
+        /// <summary>
+        /// Delete vault by item selection on vault list.
+        /// </summary>
+        /// <param name="listView"></param>
+        /// <param name="vaultDirectory"></param>
         public static void DeleteVaultItem(ListView listView, string vaultDirectory)
         {
             string vault = GetVaultNameFromListView(listView);
             if (vault.Length > 0)
             {
-                Globals.vaultName = vault;
-                //TODO
-                //DeleteVault deleteVault = new DeleteVault();
-                //deleteVault.ShowDialog();
-                if (Globals.deleteConfirmation)
+                PwMLib.GlobalVariables.vaultName = vault;
+                DeleteVault deleteVault = new DeleteVault();
+                deleteVault.ShowDialog();
+                if (PwMLib.GlobalVariables.deleteConfirmation)
                 {
                     DeleteVault(vault, vaultDirectory, listView);
-                    UtilityFunctions.VariablesClear();
+                    ClearVariables.VariablesClear();
                 }
             }
         }
+
+        /// <summary>
+        /// Get application name from selected item in listview.
+        /// </summary>
+        /// <param name="listView"></param>
+        /// <returns></returns>
 
         public static string GetVaultNameFromListView(ListView listView)
         {
             string application = string.Empty;
             if (listView.SelectedItem == null)
             {
-                Notification.ShowNotificationInfo("orange", "Warning GetVaultNameFromListView: You must select a vault to delete!");
+                Notification.ShowNotificationInfo("orange", "You must select a vault to delete!");
                 return application;
             }
             string selectedItem = listView.SelectedItem.ToString();
@@ -81,20 +93,31 @@ namespace PwM_UI.Utility
             return application;
         }
 
+        /// <summary>
+        /// Deletes a specific vault file.
+        /// </summary>
+        /// <param name="vaultName">Vault Name.</param>
+        /// <param name="password">Master password.</param>
         private static void DeleteVault(string vaultName, string vaultDirectory, ListView vaultsList)
         {
             try
             {
-                string pathToVault = Path.Combine(Globals.passwordManagerDirectory, $"{vaultName}.x");
-                if (!File.Exists(pathToVault))
+                if (vaultDirectory.StartsWith("Local"))
                 {
-                    Notification.ShowNotificationInfo("orange", $"Warning DeleteVault: Vault {vaultName} does not exist!");
+                    string pathToVault = Path.Combine(PwMLib.GlobalVariables.passwordManagerDirectory, $"{vaultName}.x");
+                    if (!File.Exists(pathToVault))
+                    {
+                        Notification.ShowNotificationInfo("orange", $"Vault {vaultName} does not exist!");
+                        return;
+                    }
+                    File.Delete(pathToVault);
+                    Notification.ShowNotificationInfo("green", $"Vault {vaultName} was deleted!");
+                    ListVaults(PwMLib.GlobalVariables.passwordManagerDirectory, vaultsList, false);
                     return;
                 }
-                File.Delete(pathToVault);
-                Notification.ShowNotificationInfo("green", $"Vault {vaultName} was deleted!");
-                ListVaults(Globals.passwordManagerDirectory, vaultsList, false);
-                return;
+                JsonManage.DeleteJsonData<VaultDetails>(PwMLib.GlobalVariables.jsonSharedVaults, f => f.Where(t => t.VaultName == vaultName + ".x" && t.SharedPath == vaultDirectory));
+                Notification.ShowNotificationInfo("green", $"Shared vault {vaultName} was removed from list!");
+                ListVaults(vaultDirectory, vaultsList, true);
             }
             catch (Exception e)
             {
@@ -102,13 +125,17 @@ namespace PwM_UI.Utility
             }
         }
 
+        /// <summary>
+        /// List current vaults in listView object.
+        /// </summary>
+        /// <param name="vaultsDirectory">Path to vault directory.</param>
         public static void ListVaults(string vaultsDirectory, ListView listView, bool enableShare)
         {
-            Globals.vaultsCount = 0;
+            PwMLib.GlobalVariables.vaultsCount = 0;
             listView.Items.Clear();
 
             if (enableShare)
-                vaultsDirectory = Globals.passwordManagerDirectory;
+                vaultsDirectory = PwMLib.GlobalVariables.passwordManagerDirectory;
 
             if (!Directory.Exists(vaultsDirectory))
             {
@@ -119,15 +146,43 @@ namespace PwM_UI.Utility
             var getFiles = new DirectoryInfo(vaultsDirectory).GetFiles();
             foreach (var file in getFiles)
             {
-                Globals.vaultsCount++;
+                PwMLib.GlobalVariables.vaultsCount++;
                 if (file.Name.EndsWith(".x"))
                 {
                     listView.Items.Add(new { Name = file.Name.Substring(0, file.Name.Length - 2), CreateDate = file.CreationTime, SharePoint = "Local Stored" });
                 }
             }
+            if (File.Exists(PwMLib.GlobalVariables.jsonSharedVaults))
+            {
+                VaultDetails[] items;
+                try
+                {
+                    items = JsonManage.ReadJsonFromFile<VaultDetails[]>(PwMLib.GlobalVariables.jsonSharedVaults);
+                    FileInfo fileInfo;
+                    foreach (var item in items)
+                    {
+                        string vaultPathFile = Path.Combine(item.SharedPath, item.VaultName);
+                        fileInfo = new FileInfo(vaultPathFile);
+                        listView.Items.Add(new { Name = item.VaultName.Substring(0, item.VaultName.Length - 2), CreateDate = fileInfo.CreationTime, SharePoint = item.SharedPath });
+                    }
+                }
+                catch
+                {
+                    Notification.ShowNotificationInfo("red", "Shared vault list is corrupted. Try import again the shared vaults!");
+                    File.Delete(PwMLib.GlobalVariables.jsonSharedVaults);
+                    return;
+                }
+            }
         }
 
-        // Clears app list, password/text boxes from application tab, then closes it and moves to the vault tab.
+        /// <summary>
+        /// Function for clear applist, and all passwords boxes and text boxes from application tab, closes it and moves to vault tab.
+        /// </summary>
+        /// <param name="homeListView"></param>
+        /// <param name="vaultListView"></param>
+        /// <param name="appListView"></param>
+        /// <param name="appList"></param>
+        /// <param name="tabControl"></param>
         public static void VaultClose(ListViewItem vaultListView, ListViewItem appListView, ListViewItem settingsListView,
             ListView appList, TabControl tabControl, DispatcherTimer masterPasswordTimer)
         {
@@ -138,12 +193,18 @@ namespace PwM_UI.Utility
             tabControl.SelectedIndex = 0;
             appListView.Foreground = Brushes.Red;
             appListView.IsEnabled = false;
-            Globals.masterPassword = null;
-            Globals.vaultOpen = false;
+            PwMLib.GlobalVariables.masterPassword = null;
+            PwMLib.GlobalVariables.vaultOpen = false;
             AppManagement.vaultSecure = null;
+            MasterPasswordTimerStart.MasterPasswordCheck_TimerStop(masterPasswordTimer);
             GC.Collect();
         }
 
+        /// <summary>
+        /// Get vault path from vault list.
+        /// </summary>
+        /// <param name="listView"></param>
+        /// <returns></returns>
         public static string GetVaultPathFromList(ListView listView)
         {
             string vaultPath = string.Empty;
@@ -155,36 +216,48 @@ namespace PwM_UI.Utility
             }
             return vaultPath;
         }
-
+        /// <summary>
+        /// Change Master Password for a vault!
+        /// </summary>
+        /// <param name="vaultList"></param>
+        /// <param name="oldMasterPassword"></param>
+        /// <param name="newMasterPassword"></param>
         public static void ChangeMassterPassword(ListView vaultList)
         {
-            string oldMasterPassword = PasswordService.SS2S(Globals.masterPassword);
-            string newMasterPassword = PasswordService.SS2S(Globals.newMasterPassword);
+            string oldMasterPassword = PasswordValidator.ConvertSecureStringToString(PwMLib.GlobalVariables.masterPassword);
+            string newMasterPassword = PasswordValidator.ConvertSecureStringToString(PwMLib.GlobalVariables.newMasterPassword);
             if (vaultList.SelectedItem == null)
             {
-                Notification.ShowNotificationInfo("orange", "Warning: You must select a vault to change your Master Password");
+                Notification.ShowNotificationInfo("orange", "You must select a vault for changeing the Master Password!");
                 return;
             }
             string vaultName = GetVaultNameFromListView(vaultList);
-            string pathToVault = pathToVault = Path.Combine(Globals.passwordManagerDirectory, $"{vaultName}.x");
+            string pathToVault;
             string vaultPath = GetVaultPathFromList(vaultList);
-
+            if (vaultPath.StartsWith("Local"))
+            {
+                pathToVault = Path.Combine(PwMLib.GlobalVariables.passwordManagerDirectory, $"{vaultName}.x");
+            }
+            else
+            {
+                pathToVault = Path.Combine(vaultPath, $"{vaultName}.x");
+            }
             if (!File.Exists(pathToVault))
             {
-                Notification.ShowNotificationInfo("red", $"Error: Vault {vaultName} does not exist!");
+                Notification.ShowNotificationInfo("red", $"Vault {vaultName} does not exist!");
                 return;
             }
             string readVault = File.ReadAllText(pathToVault);
-            string decryptVault = AesHelper.Decrypt(readVault, oldMasterPassword);
+            string decryptVault = AES.Decrypt(readVault, oldMasterPassword);
             if (decryptVault.Contains("Error decrypting"))
             {
-                Notification.ShowNotificationInfo("red", "Error: Master password invalid (or issue with vault)");
+                Notification.ShowNotificationInfo("red", "Something went wrong. Master password is incorrect or vault issue!");
                 return;
             }
-            string encryptdata = AesHelper.Encrypt(decryptVault, newMasterPassword);
+            string encryptdata = AES.Encrypt(decryptVault, newMasterPassword);
             if (!File.Exists(pathToVault))
             {
-                Notification.ShowNotificationInfo("red", $"Error: Vault {vaultName} does not exist!");
+                Notification.ShowNotificationInfo("red", $"Vault {vaultName} does not exist!");
                 return;
             }
             try
@@ -193,10 +266,8 @@ namespace PwM_UI.Utility
             }
             catch (UnauthorizedAccessException)
             {
-                Notification.ShowNotificationInfo("red", $"Error: Unauthorized.");
+                Notification.ShowNotificationInfo("red", $"Access denied: Vault is write protected for this user.");
                 return;
             }
             Notification.ShowNotificationInfo("green", $"New Master Password was set for {vaultName} vault!");
         }
-    }
-}
